Extract method + fix keyboard loading
The keyboard was being loaded in a different way when requested of
an existing app, skipping the setup that handled initial key states.
Now the thorough loading method is extracted and used in both
flows.

We still have the issue that keys initially Down (but not LockedDown)
get released when a keyboard is loaded directly into a new instance
of OKGO. This happens because the SimulateKeyStrokes property
has to be processed _after_ setting up the keyboard, and this causes
pressed (not not locked) keys to be released