Throw error if function key not recognised
It is much more likely that someone accidentally misspells a key name rather than
wanting to press/release a whole string of characters with a KeyUp/Keydown instead
of a Text. When playing a game, typing a bumch of unexpected characters gives
unexpected results and it isn't obvious what the problem is. Now we cause an obvious
error in the case where the key name isn't recognised